feat: more comments, fix tests
