+++ lib/index.ts
@@ -4,6 +4,7 @@ export * from "./bigint-math";
 export * from "./constants";
 export * from "./contract";
 export * from "./deploy";
+export * from "./deposit";
 export * from "./dsm";
 export * from "./ec";
 export * from "./eips";
@@ -14,6 +15,7 @@ export * from "./log";
 export * from "./network";
 export * from "./nor";
 export * from "./oracle";
+export * from "./pdg";
 export * from "./promise";
 export * from "./proxy";
 export * from "./scratch";
@@ -22,5 +24,3 @@ export * from "./state-file";
 export * from "./string";
 export * from "./time";
 export * from "./units";
-export * from "./deposit";
-export * from "./pdg";
