+++ test/0.8.25/vaults/delegation/contracts/VaultHub__MockForDelegation.sol
@@ -7,6 +7,8 @@ import {VaultHub} from "contracts/0.8.25/vaults/VaultHub.sol";
 import {StETH__MockForDelegation} from "./StETH__MockForDelegation.sol";
 
 contract VaultHub__MockForDelegation {
+    uint256 public constant REPORT_FRESHNESS_DELTA = 1 days;
+
     address public immutable LIDO_LOCATOR;
     StETH__MockForDelegation public immutable steth;
